requirements.txt
# general modules
django<4.3,>4.2
python-decouple
pillow

# third party modules
djangorestframework
markdown
django-filter
drf-yasg[validation]

djangorestframework-simplejwt[crypto]
coreapi
# deployment modules


# third party email modules
django-mail-templated

# reformat and lint
flake8
black